feat: se agrega sucursales

+++ aura-pos/src/main/java/com/cloud_technological/aura_pos/repositories/sucursales/SucursalJPARepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.cloud_technological.aura_pos.entity.SucursalEntity;
 
 public interface SucursalJPARepository extends JpaRepository<SucursalEntity, Integer> {
-    Optional<SucursalEntity> findByIdAndEmpresaId(Long id, Integer empresaId);
+    Optional<SucursalEntity> findByIdAndEmpresaId(Integer id, Integer empresaId);
 }
