+++ aura-pos/src/main/java/com/cloud_technological/aura_pos/services/implementations/SerialProductoServiceImpl.java
@@ -72,7 +72,7 @@ public class SerialProductoServiceImpl implements SerialProductoService {
         ProductoEntity producto = productoJPARepository.findByIdAndEmpresaId(dto.getProductoId(), empresaId)
                 .orElseThrow(() -> new GlobalException(HttpStatus.BAD_REQUEST, "Producto no encontrado"));
 
-        SucursalEntity sucursal = sucursalJPARepository.findByIdAndEmpresaId(dto.getSucursalId(), empresaId)
+        SucursalEntity sucursal = sucursalJPARepository.findByIdAndEmpresaId(dto.getSucursalId().intValue(), empresaId)
                 .orElseThrow(() -> new GlobalException(HttpStatus.BAD_REQUEST, "Sucursal no encontrada"));
 
         SerialProductoEntity entity = serialMapper.toEntity(dto);
